Move into consistent folder structure, remove bad cleanupcode eof linebreaks
