+++ test/OpenApiClientTests/SchemaProperties/NullableReferenceTypesDisabled/swagger.g.json
@@ -1228,7 +1228,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1297,7 +1297,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
@@ -1334,7 +1334,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1357,7 +1357,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1404,7 +1404,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1433,7 +1433,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1459,7 +1459,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1528,7 +1528,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
@@ -1569,7 +1569,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1664,7 +1664,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
@@ -1821,7 +1821,7 @@
               "type": "array"
             }
           ],
-          "items": {}
+          "items": { }
         },
         "nullable": true
       },
@@ -1844,7 +1844,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1874,7 +1874,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           },
           "jsonapi": {
             "$ref": "#/components/schemas/jsonapiObject"
@@ -1925,7 +1925,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
@@ -1962,7 +1962,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
@@ -1993,7 +1993,7 @@
           },
           "meta": {
             "type": "object",
-            "additionalProperties": {}
+            "additionalProperties": { }
           }
         },
         "additionalProperties": false
