+++ test/OpenApiTests/SchemaProperties/NullableReferenceTypesDisabled/NullabilityTests.cs
@@ -22,7 +22,7 @@ public sealed class NullabilityTests
     [Theory]
     [InlineData("name")]
     [InlineData("timeAtCurrentFarmInDays")]
-    public async Task Property_in_schema_for_resource_should_be_nullable(string propertyName)
+    public async Task Property_in_schema_for_attribute_of_resource_should_be_nullable(string propertyName)
     {
         // Act
         JsonElement document = await _testContext.GetSwaggerDocumentAsync();
@@ -42,7 +42,7 @@ public sealed class NullabilityTests
     [InlineData("age")]
     [InlineData("weight")]
     [InlineData("hasProducedEggs")]
-    public async Task Property_in_schema_for_resource_should_not_be_nullable(string propertyName)
+    public async Task Property_in_schema_for_attribute_of_should_not_be_nullable(string propertyName)
     {
         // Act
         JsonElement document = await _testContext.GetSwaggerDocumentAsync();
@@ -56,5 +56,40 @@ public sealed class NullabilityTests
             });
         });
     }
-}
 
+    [Theory]
+    [InlineData("oldestChicken")]
+    public async Task Property_in_schema_for_relationship_of_resource_should_be_nullable(string propertyName)
+    {
+        // Act
+        JsonElement document = await _testContext.GetSwaggerDocumentAsync();
+
+        // Assert
+        document.ShouldContainPath("components.schemas.henHouseRelationshipsInPostRequest.properties").With(schemaProperties =>
+        {
+            schemaProperties.ShouldContainPath($"{propertyName}.$ref").WithSchemaReferenceId(schemaReferenceId =>
+            {
+                document.ShouldContainPath($"components.schemas.{schemaReferenceId}.properties.data.oneOf[1].$ref").ShouldBeSchemaReferenceId("nullValue");
+            });
+        });
+    }
+
+    [Theory]
+    [InlineData("allChickens")]
+    [InlineData("firstChicken")]
+    [InlineData("chickensReadyForLaying")]
+    public async Task Data_property_in_schema_for_relationship_of_resource_should_not_be_nullable(string propertyName)
+    {
+        // Act
+        JsonElement document = await _testContext.GetSwaggerDocumentAsync();
+
+        // Assert
+        document.ShouldContainPath("components.schemas.henHouseRelationshipsInPostRequest.properties").With(schemaProperties =>
+        {
+            schemaProperties.ShouldContainPath($"{propertyName}.$ref").WithSchemaReferenceId(schemaReferenceId =>
+            {
+                document.ShouldContainPath($"components.schemas.{schemaReferenceId}.properties.data").ShouldNotContainPath("oneOf[1].$ref");
+            });
+        });
+    }
+}
