+++ test/OpenApiTests/SchemaProperties/NullableReferenceTypesEnabled/Cow.cs
@@ -37,5 +37,3 @@ public sealed class Cow : Identifiable<int>
     [Required]
     public bool? HasProducedMilk { get; set; }
 }
-
-
